MediatR, Commands, Handlers , Queries have been added

+++ ITEPortal.Domain/Dto/AuthCodeDto.cs
@@ -2,10 +2,10 @@
 {
     public class AuthCodeDto
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string CodeNumber { get; set; } = string.Empty;
         public DateTime ExpiredDate { get; set; }
 
-        public long UserId { get; set; }
+        public int UserId { get; set; }
     }
 }
