+++ ITEPortal.Domain/Dto/UserDto.cs
@@ -6,12 +6,12 @@
         /// The Id of User
         /// </summary>
         /// <example>1</example>
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         public string FirstName { get; set; } = string.Empty;
         public string LastName { get; set; } = string.Empty;
         public string Email { get; set; } = string.Empty;
         public string PhoneNumber { get; set; } = string.Empty;
-        public long UserRoleId { get; set; }
+        public int UserRoleId { get; set; }
     }
 }
